+++ src/jtl/Connector/Presta/Controller/Category.php
@@ -10,7 +10,7 @@ class Category extends BaseController
 		$result = $this->db->executeS('
 			SELECT c.* 
 			FROM '._DB_PREFIX_.'category c
-			LEFT JOIN jtl_connector_link l ON c.id_category = l.endpointId AND l.type = 1
+			LEFT JOIN jtl_connector_link_category l ON c.id_category = l.endpointId AND l.type = 1
             WHERE l.hostId IS NULL AND c.id_parent != 0 AND c.is_root_category = 0
             ORDER BY c.nleft
             LIMIT '.$limit
@@ -61,7 +61,7 @@ class Category extends BaseController
 		return $this->db->getValue('
 			SELECT COUNT(*) 
 			FROM '._DB_PREFIX_.'category c
-			LEFT JOIN jtl_connector_link l ON c.id_category = l.endpointId AND l.type = 1
+			LEFT JOIN jtl_connector_link_category l ON c.id_category = l.endpointId AND l.type = 1
             WHERE l.hostId IS NULL AND c.id_parent != 0 AND c.is_root_category = 0
         ');
 	}