+++ src/jtl/Connector/Presta/Controller/Customer.php
@@ -10,7 +10,7 @@ class Customer extends BaseController
 			FROM '._DB_PREFIX_.'customer c
 			LEFT JOIN '._DB_PREFIX_.'address a ON a.id_customer=c.id_customer
 			LEFT JOIN '._DB_PREFIX_.'country co ON co.id_country=a.id_country
-			LEFT JOIN jtl_connector_link l ON c.id_customer = l.endpointId AND l.type = 2
+			LEFT JOIN jtl_connector_link_customer l ON c.id_customer = l.endpointId AND l.type = 2
             WHERE l.hostId IS NULL
             GROUP BY c.id_customer
             LIMIT '.$limit
@@ -47,7 +47,7 @@ class Customer extends BaseController
 		return $this->db->getValue('
 			SELECT COUNT(*) 
 			FROM '._DB_PREFIX_.'customer c
-			LEFT JOIN jtl_connector_link l ON c.id_customer = l.endpointId AND l.type = 2
+			LEFT JOIN jtl_connector_link_customer l ON c.id_customer = l.endpointId AND l.type = 2
             WHERE l.hostId IS NULL
         ');
 	}