+++ src/jtl/Connector/Presta/Controller/Image.php
@@ -181,7 +181,7 @@ class Image extends BaseController
     {
         $categories = $this->db->executeS('
           SELECT c.id_category FROM '._DB_PREFIX_.'category c
-          LEFT JOIN jtl_connector_link l ON CONCAT("c", c.id_category) = l.endpointId AND l.type = 16
+          LEFT JOIN jtl_connector_link_image l ON CONCAT("c", c.id_category) = l.endpointId AND l.type = 16
           WHERE l.hostId IS NULL
         ');
 
@@ -206,7 +206,7 @@ class Image extends BaseController
     {
         $manufacturers = $this->db->executeS('
           SELECT m.id_manufacturer FROM '._DB_PREFIX_.'manufacturer m
-          LEFT JOIN jtl_connector_link l ON CONCAT("m", m.id_manufacturer) = l.endpointId AND l.type = 16
+          LEFT JOIN jtl_connector_link_image l ON CONCAT("m", m.id_manufacturer) = l.endpointId AND l.type = 16
           WHERE l.hostId IS NULL
         ');
 
@@ -231,7 +231,7 @@ class Image extends BaseController
     {
         $images = $this->db->executeS('
           SELECT i.* FROM '._DB_PREFIX_.'image i
-          LEFT JOIN jtl_connector_link l ON i.id_image = l.endpointId AND l.type = 16
+          LEFT JOIN jtl_connector_link_image l ON i.id_image = l.endpointId AND l.type = 16
           WHERE l.hostId IS NULL
         ');