+++ src/jtl/Connector/Presta/Controller/Manufacturer.php
@@ -8,7 +8,7 @@ class Manufacturer extends BaseController
 		$result = $this->db->executeS('
 			SELECT m.*
 			FROM '._DB_PREFIX_.'manufacturer m
-			LEFT JOIN jtl_connector_link l ON m.id_manufacturer = l.endpointId AND l.type = 32
+			LEFT JOIN jtl_connector_link_manufacturer l ON m.id_manufacturer = l.endpointId AND l.type = 32
             WHERE l.hostId IS NULL
             LIMIT '.$limit
         );
@@ -55,7 +55,7 @@ class Manufacturer extends BaseController
 		return $this->db->getValue('
 			SELECT COUNT(*) 
 			FROM '._DB_PREFIX_.'manufacturer m
-			LEFT JOIN jtl_connector_link l ON m.id_manufacturer = l.endpointId AND l.type = 32
+			LEFT JOIN jtl_connector_link_manufacturer l ON m.id_manufacturer = l.endpointId AND l.type = 32
             WHERE l.hostId IS NULL
         ');
 	}