+++ changelog.md
@@ -1,6 +1,7 @@
 1.6.0.1
 ------
 - CO-552 Ignore product-images which are not assigned to a product
+- Linking table access is now fixed
 
 1.5.9
 ------